convert barloader to typescript
