+++ src/BarLoader.tsx
@@ -1,0 +1,72 @@
+/** @jsx jsx */
+import * as React from "react";
+import { keyframes, css, jsx } from "@emotion/core";
+import onlyUpdateForKeys from "recompose/onlyUpdateForKeys";
+import { calculateRgba, heightWidthDefaults, heightWidthKeys } from "./helpers";
+import { Keyframes } from "@emotion/serialize";
+import { BarLoaderProps, StyleFunction, PrecompiledCss } from "./interfaces";
+
+const long: Keyframes = keyframes`
+  0% {left: -35%;right: 100%} 
+  60% {left: 100%;right: -90%}
+  100% {left: 100%;right: -90%}
+`;
+
+const short: Keyframes = keyframes`
+  0% {left: -200%;right: 100%} 
+  60% {left: 107%;right: -8%}
+  100% {left: 107%;right: -8%}
+`;
+
+export class Loader extends React.PureComponent<BarLoaderProps> {
+  static defaultProps: BarLoaderProps = heightWidthDefaults(4, 100);
+
+  style: StyleFunction = (i: number): PrecompiledCss => {
+    const { height, color, heightUnit } = this.props;
+
+    return css`
+      position: absolute;
+      height: ${`${height}${heightUnit}`};
+      overflow: hidden;
+      background-color: ${color};
+      background-clip: padding-box;
+      display: block;
+      border-radius: 2px;
+      will-change: left, right;
+      animation-fill-mode: forwards;
+      animation: ${i === 1 ? long : short} 2.1s ${i === 2 ? "1.15s" : ""}
+        ${i === 1
+          ? "cubic-bezier(0.65, 0.815, 0.735, 0.395)"
+          : "cubic-bezier(0.165, 0.84, 0.44, 1)"}
+        infinite;
+    `;
+  };
+
+  wrapper: StyleFunction = (): PrecompiledCss => {
+    const { width, height, color, heightUnit, widthUnit } = this.props;
+
+    return css`
+      position: relative;
+      width: ${`${width}${widthUnit}`};
+      height: ${`${height}${heightUnit}`};
+      overflow: hidden;
+      background-color: ${calculateRgba(color, 0.2)};
+      background-clip: padding-box;
+    `;
+  };
+
+  render(): JSX.Element | null {
+    const { loading, css } = this.props;
+
+    return loading ? (
+      <div css={[this.wrapper(), css]}>
+        <div css={this.style(1)} />
+        <div css={this.style(2)} />
+      </div>
+    ) : null;
+  }
+}
+
+const Component = onlyUpdateForKeys(heightWidthKeys)(Loader);
+Component.defaultProps = Loader.defaultProps;
+export default Component;
